requirements.txt
# Core data processing - Python 3.12 compatible versions
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3

# Excel handling
openpyxl==3.1.2
xlrd==2.0.1

# Web requests
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# File monitoring
watchdog==4.0.0

# Environment variables
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0